requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
statsmodels
tensorflow-cpu
prophet
cmdstanpy
deep-translator
openai
requests
scipy
sentence-transformers
